requirements.txt
alpha-vantage==2.0.0
amqp==2.3.2
beautifulsoup4==4.6.0
billiard==3.5.0.4
celery==4.2.1
certifi==2018.4.16
chardet==3.0.4
cycler==0.10.0
decorator==4.3.0
Django==2.0.7
gurobipy==7.5.2
idna==2.7
iexfinance==0.3.4
inflect==1.0.0
inflection==0.3.1
ipython-genutils==0.2.0
jaraco.itertools==2.4
jsonschema==2.6.0
jupyter-core==4.4.0
kiwisolver==1.0.1
kombu==4.2.1
lxml==4.2.3
matplotlib==2.2.2
more-itertools==4.2.0
nbformat==4.4.0
numpy==1.14.5
pandas==0.23.4
pandas-datareader==0.6.0
patsy==0.5.0
plotly==3.0.2
psycopg2==2.7.5
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
Quandl==3.4.0
redis==2.10.6
requests==2.19.1
requests-file==1.4.3
requests-ftp==0.3.1
retrying==1.3.3
scipy==1.1.0
selenium==3.14.0
simplejson==3.16.0
six==1.11.0
statsmodels==0.9.0
traitlets==4.3.2
urllib3==1.23
vine==1.1.4
wrapt==1.10.11
xmltodict==0.11.0
